+++ Sprint_2/code/frontend/pages/myRoute/index.vue
@@ -26,6 +26,7 @@ Contributer: Chetsada
 - เพิ่มส่วนฟิลเตอร์กรองรีวิวตามดาว const selectedStarFilter เป็น state เก็บดาว
 - เพิ่ม const filteredReviews สำหรับกรองดาว
 - เพิ่ม const starCounts สำหรับนับจำนวนรีวิวในดาวนั้นๆ
+- แก้การแสดงผลคะแนนรีวิว ที่ Driver Profile เพิ่ม .toFixed(1) ในส่วนตกหล่น ให้แสดงจุดทศนิยมแค่จุดเดียว
 [AI Declare]
 - ใช้ ChatGPT ออกแบบฟังก์ชัน 
 -->
@@ -445,7 +446,7 @@ Contributer: Chetsada
                       </button>
                     </div>
 
-                    <!--reviewpopup-->
+                    <!-- Driver Profile -->
                     <div class="flex items-center mt-1 cursor-pointer" @click.stop="openReviewModal(trip)">
                       <div class="flex text-sm text-yellow-400">
                         <span>
@@ -456,10 +457,8 @@ Contributer: Chetsada
                         </span>
                       </div>
                       <span class="ml-2 text-sm text-gray-600">
-                        {{ trip.passenger.rating }} ({{
-                          trip.passenger.reviews
-                        }}
-                        รีวิว)
+                        {{ (trip.passenger.rating ?? 0).toFixed(1)  }} 
+                        ({{ trip.passenger.reviews ?? 0}} รีวิว)
                       </span>
                     </div>
 
@@ -687,8 +686,8 @@ Contributer: Chetsada
                   </span>
                 </div>
                 <span class="ml-2 text-sm text-gray-600">
-                  {{ (driverInfo?.rating || 0).toFixed(1) }} 
-                  ({{ driverInfo?.reviews || 0 }} รีวิว)
+                  {{ (driverInfo?.rating ?? 0).toFixed(1) }} 
+                  ({{ driverInfo?.reviews ?? 0 }} รีวิว)
                 </span>
               </div>
             </div>