+++ Sprint_2/doc/Change log.md
@@ -61,6 +61,9 @@ This file will contain all notable changes to this project.
       - เพิ่มส่วนฟิลเตอร์กรองรีวิวตามดาว const selectedStarFilter เป็น state เก็บดาว
       - เพิ่ม const filteredReviews สำหรับกรองดาว
       - เพิ่ม const starCounts สำหรับนับจำนวนรีวิวในดาวนั้นๆ
+- ที่ ..frontend\pages\myTrip\index.vue
+    ..frontend\pages\myRoute\index.vue
+      - แก้การแสดงผลคะแนนรีวิว ที่ Driver Profile เพิ่ม .toFixed(1) ในส่วนตกหล่น ให้แสดงจุดทศนิยมแค่จุดเดียว
 ### AI Declare
 - ใช้ ChatGPT ออกแบบฟังก์ชัน และเขียน vue แสดงฟิลเตอร์ดาว
 ---